mcp/requirements.txt
pyyaml==6.0.2
requests==2.32.3
fastmcp==0.3.0
openapi-spec-validator==0.7.1
openai==1.43.0
anthropic==0.34.0
